Remove chat_format override
This is fundamentally not compatible with the way chat has changed in 1.19, as message decoration with player name is handled on the client. The most common use-case is replaced by the previously added name_decoration override, and the remaining case will be covered by an override to specify a data-driven chat type to use for all players with a role.